rename PageHeader to Header and Header to Sidebar
