+++ app/_components/organisms/index.ts
@@ -1,21 +1,21 @@
 import Wrapper from "./Wrapper";
-import Header from "./Header";
+import Sidebar from "./Sidebar";
 import Footer from "./Footer";
 import Loading from "./Loading";
 import Navigation from "./Navigation";
 import User from "./User";
-import PageHeader from "./PageHeader";
+import Header from "./Header";
 import Content from "./Content";
 import Authenticate from "./Authenticate";
 
 const Organism = {
   Wrapper,
-  Header,
+  Sidebar,
   Footer,
   Loading,
   Navigation,
   User,
-  PageHeader,
+  Header,
   Authenticate,
   Content,
 }
