+++ app/modules/page.tsx
@@ -4,9 +4,9 @@ import { FlagOutlined } from "@ant-design/icons";
 export default async function ModulesPage() {
   return (
     <div>
-      <Organism.PageHeader title="Modules" icon={<FlagOutlined style={{ fontSize: "20px"}} />}>
+      <Organism.Header title="Modules" icon={<FlagOutlined style={{ fontSize: "20px"}} />}>
         Menu Here
-      </Organism.PageHeader>
+      </Organism.Header>
       <Organism.Content>
         Content
       </Organism.Content>
